requirements.txt
fastapi
uvicorn
dlib
face_recognition
faiss-cpu
numpy
python-multipart
boto3